Add several model class definitions & migrations
Add NewPostEmail, ClaimedItemEmail, EmailThread, and Item to models.py,
along with the database migration (0002) to support these classes. This
should allow work to be done on the public API based off of these models.

In addition, this commit restructures the test package for the web_api
Django app, and adds a modest set of unit tests for the afforementioned
model classes.