Add modified field to EmailThread
The save() methods on Item, NewPostEmail, and ClaimedItemEmail have been
overriden to automatically update the EmailThread.modified field. In terms
of efficiency, this makes more sense than having a property on EmailThread
(implemented by overriding __getatr__) that pulls the values of the
associated Item, NewPostEmail, and ClaimedEmail objects. The reasoning
here is that the field will be read much more often than written, so
making it usable in an SQL query condition is more important than
preventing a little bit of data duplication.