Add SMTP incoming email processing
The mechanism uses two pieces of code: smtp.IncomingServer and
smtp.EmailReceiver. The IncomingServer class is run as root, and is
effectively a relay from port 25 to port 7999, which can be connected to
by user programs. The EmailReceiver class runs as a client that connects
to port 7999 and waits for emails relayed from the IncomingServer server.

The EmailReceiver then uses EmailParser and company to process incoming
emails and insert them into the database.